feat: 拆分函数
